[Refact] Read update event 경로 수정
