+++ src/main/java/com/chat_server/chatread/service/ChatReadService.java
@@ -13,4 +13,14 @@ public interface ChatReadService {
      * @return 읽음 갱신 이벤트 DTO
      */
     ChatReadUpdatedEvent read(ChatReadRequest request, Long userId);
+
+    /**
+     * 채팅방 입장 시 최신 메시지 ID 기준으로 읽음 상태를 갱신한다.
+     *
+     * @param roomId 채팅방 ID
+     * @param userId 사용자 ID
+     * @param latestMessageId 채팅방의 최신 메시지 ID. 메시지가 없으면 null
+     */
+    void markAsReadOnEnter(Long roomId, Long userId, Long latestMessageId);
+
 }
