--- src/main/java/com/chat_server/chatroom/service/impl/ChatReadServiceImpl.java
@@ -1,42 +1,0 @@
-package com.chat_server.chatroom.service.impl;
-
-import com.chat_server.chatlist.service.ChatListService;
-import com.chat_server.chatroom.service.ChatReadService;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-@Slf4j
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class ChatReadServiceImpl implements ChatReadService {
-    private final ChatListService chatListService;
-
-    /**
-     * 채팅방 입장 시 unread_count를 0으로 만들고,
-     * 최신 메시지가 있으면 last_read_message_id를 해당 값까지 올린다.
-     *
-     * @param roomId 채팅방 ID
-     * @param userId 사용자 ID
-     * @param latestMessageId 최신 메시지 ID. null이면 메시지가 없는 방으로 간주한다.
-     */
-    @Override
-    public void markAsReadOnEnter(Long roomId, Long userId, Long latestMessageId) {
-        int updatedRows;
-
-        if(latestMessageId == null){
-            updatedRows = chatListService.clearUnreadCountOnEnter(roomId, userId);
-        }else{
-            updatedRows = chatListService.markAsReadOnEnter(roomId, userId, latestMessageId);
-        }
-
-        if (updatedRows == 0) {
-            throw new IllegalStateException(
-                    "chat_list 멤버십이 없어 입장 읽음 처리를 할 수 없습니다. roomId="
-                            + roomId + ", userId=" + userId
-            );
-        }
-    }
-}
